Get the first feature working (can start a game)
