+++ app/javascript/components/GameBoardSetter.js
@@ -1,30 +1,43 @@
-import React, { useState, useEffect, useRef } from "react"
+import React, { useState, useEffect } from "react"
 import PropTypes from "prop-types"
 import GameBoard from "./GameBoard";
 import { setSetters } from '../state/CardState'
+import { setCards, pollEvents } from '../state/CardState';
 
 const GameBoardSetter = (props) => {
   const [locations, setLocations] = useState(props.locations);
 
+  setCards(props.cards);
+
+  if(!props.skipPolling) {
+    useEffect(() => {
+      let interval = setInterval(pollEvents, 5000);
+      return () => {
+        clearInterval(interval)
+      }
+    })
+  }
+
   useEffect(() => {
-    setSetters({setLocations: setLocations})
+    setSetters({setLocations: (l) => {
+      debugger;
+      console.log(l);
+      setLocations({...locations, ...l})
+    }});
+    return () => { setSetters({setLocations: () => {}}) }
   });
 
-  try {
-    return <GameBoard
-      id={props.id}
-      key={props.key}
-      name={props.name}
-      gameId={props.gameId}
-      cards={props.cards}
-      locations={locations}
-      stacks={props.stacks}
-      params={props.params}
-      skipPolling={props.skipPolling}
-    />
-  } catch (e){
-    console.log(e)
-  }
+  return <GameBoard
+    id={props.id}
+    key={props.key}
+    name={props.name}
+    gameId={props.gameId}
+    cards={props.cards}
+    locations={locations}
+    stacks={props.stacks}
+    params={props.params}
+    skipPolling={props.skipPolling}
+  />
 }
 
 GameBoardSetter.propTypes = {
@@ -35,7 +48,6 @@ GameBoardSetter.propTypes = {
   cards: PropTypes.array,
   locations: PropTypes.array,
   stacks: PropTypes.object,
-  params: PropTypes.object,
-  skipPolling: PropTypes.bool,
+  params: PropTypes.object
 };
 export default GameBoardSetter
